Make DMMemoize thread-safe
This was my weekend exercise. It's a lot of extra code! The locking is
fairly complex. NSMapTable isn't thread-safe (which is sane), so we must
lock around our use of it. However, the generator block runs arbitrary
code, including calling the memoized method recursively (like the
canonical Fibonacci example), so we can't keep the NSMapTable locked
while calling the generatorBlock or we could easily deadlock.

It was also tricky to arrange for generatorBlock to be called exactly
once - that is, if the value isn't cached but another thread is actively
generating it, we wait for that thread (and release our lock while doing
so). When the generating thread finishes, it must signal all other
threads waiting on that value, so they can re-acquire the lock and read
out the now-cached value. I accomplish this with an additional
NSMapTable, mapping each in-progess key to an empty semaphore.